интеракции, оверлей, shared ерп панель
