+++ Resources/Locale/en-US/_Arcane/ERP/erp-panel.ftl
@@ -43,28 +43,28 @@ orgasm-message-5 = quietly moans while finishing!
 orgasm-message-6 = moans loudly, reaching the peak of pleasure!
 
 # Interaction names
-erp-panel-interaction-anus-fuck-name = Fuck anus
-erp-panel-interaction-anus-lick-name = Lick anus
-erp-panel-interaction-anus-fingering-name = Finger anus
-erp-panel-interaction-anus-tail-fuck-name = Tail-fuck anus
-erp-panel-interaction-anal-lick2-name = Deeply lick anus
-erp-panel-interaction-anal-fingering2-name = Use two fingers
-erp-panel-interaction-toy-butt-name = Use toy
-erp-panel-interaction-boobs-slap-name = Slap breasts
-erp-panel-interaction-boobs-paw-name = Stroke breasts
-erp-panel-interaction-boobs-fuck-name = Fuck breasts
-erp-panel-interaction-boobs-squeeze-name = Squeeze breasts
-erp-panel-interaction-boobs-kisses-name = Kiss breasts
-erp-panel-interaction-boobs-tail-name = Caress breasts with tail
-erp-panel-interaction-boobs-massage-name = Massage breasts
+erp-panel-interaction-anus-fuck-name = Fuck ass
+erp-panel-interaction-anus-lick-name = Lick ass
+erp-panel-interaction-anus-fingering-name = Finger ass
+erp-panel-interaction-anus-tail-fuck-name = Tail-fuck ass
+erp-panel-interaction-anal-lick2-name = Lick ass deep
+erp-panel-interaction-anal-fingering2-name = Double-finger ass
+erp-panel-interaction-toy-butt-name = Insert toy
+erp-panel-interaction-boobs-slap-name = Slap tits
+erp-panel-interaction-boobs-paw-name = Fondle tits
+erp-panel-interaction-boobs-fuck-name = Fuck tits
+erp-panel-interaction-boobs-squeeze-name = Squeeze tits
+erp-panel-interaction-boobs-kisses-name = Kiss tits
+erp-panel-interaction-boobs-tail-name = Tail-fondle tits
+erp-panel-interaction-boobs-massage-name = Knead tits
 erp-panel-interaction-boobs-suck-name = Suck nipple
-erp-panel-interaction-butt-near-cock-name = Rub cock against buttocks
-erp-panel-interaction-butt-massage-name = Massage butt
-erp-panel-interaction-butt-tail-name = Stroke butt with tail
+erp-panel-interaction-butt-near-cock-name = Rub cock on ass
+erp-panel-interaction-butt-massage-name = Knead ass
+erp-panel-interaction-butt-tail-name = Tail-fondle butt
 erp-panel-interaction-hands-on-butt-name = Grab butt
 erp-panel-interaction-kiss-butt-name = Kiss butt
 erp-panel-interaction-slap-butt-name = Slap butt
-erp-panel-interaction-horns-ass-f-name = Rub horn against butt
+erp-panel-interaction-horns-ass-f-name = Rub horn on butt
 erp-panel-category-face-name = Face
 erp-panel-category-butt-name = Butt
 erp-panel-category-boobs-name = Breasts
@@ -84,20 +84,20 @@ erp-panel-interaction-bite-ears-name = Bite ear
 erp-panel-interaction-bite-ears-human-name = Bite ear
 erp-panel-interaction-ears-kiss-name = Kiss ear
 erp-panel-interaction-pat-ears-name = Pet ears
-erp-panel-interaction-pat-ears-human-name = Play with ears
+erp-panel-interaction-pat-ears-human-name = Tug ears
 erp-panel-interaction-cheeks-kiss-name = Kiss cheek
 erp-panel-interaction-cheeks-kiss-both-name = Kiss both cheeks
 erp-panel-interaction-cheeks-pat-name = Pat cheeks
 erp-panel-interaction-cheeks-press-name = Squeeze cheeks
-erp-panel-interaction-gentle-kiss-name = Gentle kiss
-erp-panel-interaction-kiss-lips-name = Passionate kiss
-erp-panel-interaction-face-rub-cat-name = Rub muzzle against face
+erp-panel-interaction-gentle-kiss-name = Kiss gently
+erp-panel-interaction-kiss-lips-name = Make out
+erp-panel-interaction-face-rub-cat-name = Rub muzzles
 erp-panel-interaction-face-sitting-name = Sit on face
 erp-panel-interaction-face-dick-slap-name = Slap face with cock
-erp-panel-interaction-legs-squeezeface-name = Squeeze face with thighs
-erp-panel-interaction-mewing-name = Meow into face
+erp-panel-interaction-legs-squeezeface-name = Thigh-squeeze face
+erp-panel-interaction-mewing-name = Meow in face
 erp-panel-interaction-handshake-name = Shake hands
-erp-panel-interaction-finger-tap-name = Massage back
+erp-panel-interaction-finger-tap-name = Knead back
 erp-panel-interaction-tail-pat-name = Pet tail
 erp-panel-interaction-tail-hug-name = Entwine tails
 erp-panel-interaction-high-five-name = High five
@@ -107,33 +107,33 @@ erp-panel-interaction-friendly-tail-pull-name = Pull tail
 erp-panel-interaction-kiss-hand-name = Kiss hand
 erp-panel-interaction-pat-shoulder-name = Pat shoulder
 erp-panel-interaction-pet-head-name = Pat head
-erp-panel-interaction-playful-nibble-name = Playfully nibble
+erp-panel-interaction-playful-nibble-name = Play-bite
 erp-panel-interaction-horns-pat-name = Pet horns
-erp-panel-interaction-kitty-name = Nuzzle
-erp-panel-interaction-feet-mouth-name = Foot fetish
+erp-panel-interaction-kitty-name = Rub muzzles
+erp-panel-interaction-feet-mouth-name = Suck feet
 erp-panel-interaction-bite-neck-name = Bite neck
 erp-panel-interaction-lick-neck-name = Lick neck
 erp-panel-interaction-kiss-neck-name = Kiss neck
-erp-panel-interaction-shoulder-massage-name = Massage shoulders
+erp-panel-interaction-shoulder-massage-name = Knead shoulders
 erp-panel-interaction-nuzzle-neck-name = Nuzzle neck
-erp-panel-interaction-nuzzle-neck-vulpa-name = Nuzzle neck with muzzle
+erp-panel-interaction-nuzzle-neck-vulpa-name = Nuzzle neck
 erp-panel-interaction-blow-slap-name = Slap face
 erp-panel-interaction-suck-cock-name = Suck cock
-erp-panel-interaction-slime-fuck-cock-name = Fuck cock with slime
-erp-panel-interaction-cock-fap-name = Stroke cock
-erp-panel-interaction-cock-footjob-name = Footjob cock
+erp-panel-interaction-slime-fuck-cock-name = Slime-fuck cock
+erp-panel-interaction-cock-fap-name = Jerk cock
+erp-panel-interaction-cock-footjob-name = Footjob
 erp-panel-interaction-cock-swordmanship-name = Swordfight cocks
-erp-panel-interaction-cock-fap-tail-name = Stroke cock with tail
-erp-panel-interaction-cock-fuck-legs-name = Fuck thighs
-erp-panel-interaction-cock-blowjob-name = Give blowjob
-erp-panel-interaction-cock-near-ass-tail-name = Rub cock against tailed butt
-erp-panel-interaction-hair-face-fuck-name = Facefuck while holding hair
-erp-panel-interaction-legs-cock-fuck-tail-name = Tail-assisted thighjob
+erp-panel-interaction-cock-fap-tail-name = Tailjob
+erp-panel-interaction-cock-fuck-legs-name = Thighjob
+erp-panel-interaction-cock-blowjob-name = Give head
+erp-panel-interaction-cock-near-ass-tail-name = Rub cock on ass
+erp-panel-interaction-hair-face-fuck-name = Facefuck
+erp-panel-interaction-legs-cock-fuck-tail-name = Tail-thighjob
 erp-panel-interaction-horns-blowjob-f-name = Lick horn
-erp-panel-interaction-pussy-fingering-name = Finger vagina
-erp-panel-interaction-pussy-licking-name = Lick vagina
-erp-panel-interaction-pussy-scissors-name = Scissor
-erp-panel-interaction-pussy-touch-name = Touch vagina
-erp-panel-interaction-pussy-tail-fuck-name = Tail-fuck vagina
+erp-panel-interaction-pussy-fingering-name = Finger pussy
+erp-panel-interaction-pussy-licking-name = Lick pussy
+erp-panel-interaction-pussy-scissors-name = Scissor pussies
+erp-panel-interaction-pussy-touch-name = Fondle pussy
+erp-panel-interaction-pussy-tail-fuck-name = Tail-fuck pussy
 erp-panel-interaction-cock-pussy-fuck-name = Fuck pussy
-erp-panel-interaction-toy-pussy-name = Use toy
+erp-panel-interaction-toy-pussy-name = Insert toy
