+++ Resources/Prototypes/_Arcane/Interactions/ERP/anus.yml
@@ -3,7 +3,7 @@
   category: Anus
   userArouse: 10
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/asscrap1.ogg
   - /Audio/_Arcane/Interactions/asscrap2.ogg
@@ -35,7 +35,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -56,7 +56,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -76,7 +76,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/asscrap1.ogg
   - /Audio/_Arcane/Interactions/asscrap2.ogg
@@ -109,7 +109,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -130,7 +130,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -151,7 +151,7 @@
   category: Anus
   userArouse: 5
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/asscrap1.ogg
   - /Audio/_Arcane/Interactions/asscrap2.ogg
