+++ Resources/Prototypes/_Arcane/Interactions/ERP/boobs.yml
@@ -2,7 +2,7 @@
   id: BoobsSlap
   category: Boobs
   userArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -29,7 +29,7 @@
   id: BoobsPaw
   category: Boobs
   userArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/friction1.ogg
   - /Audio/_Arcane/Interactions/friction2.ogg
@@ -54,7 +54,7 @@
   category: Boobs
   userArouse: 10
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -90,7 +90,7 @@
   id: BoobsSqueeze
   category: Boobs
   userArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/pat1.ogg
   - /Audio/_Arcane/Interactions/pat2.ogg
@@ -116,7 +116,7 @@
   category: Boobs
   userArouse: 5
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/kiss1.ogg
   - /Audio/_Arcane/Interactions/kiss2.ogg
@@ -146,7 +146,7 @@
   id: BoobsTail
   category: Boobs
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/friction1.ogg
   - /Audio/_Arcane/Interactions/friction2.ogg
@@ -181,7 +181,7 @@
   category: Boobs
   userArouse: 5
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/pat1.ogg
   - /Audio/_Arcane/Interactions/pat2.ogg
@@ -206,7 +206,7 @@
   category: Boobs
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/oral1.ogg
   - /Audio/_Arcane/Interactions/oral2.ogg
