+++ Resources/Prototypes/_Arcane/Interactions/ERP/penis.yml
@@ -3,7 +3,7 @@
   category: Penis
   userArouse: 5
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/oral1.ogg
   - /Audio/_Arcane/Interactions/oral2.ogg
@@ -31,7 +31,7 @@
   category: Penis
   userArouse: 10
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bj1.ogg
   - /Audio/_Arcane/Interactions/bj2.ogg
@@ -69,7 +69,7 @@
   category: Penis
   userArouse: 10
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   selfMessages:
@@ -95,7 +95,7 @@
   id: CockFootjob
   category: Penis
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -125,7 +125,7 @@
   category: Penis
   userArouse: 10
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -159,7 +159,7 @@
   id: CockFapTail
   category: Penis
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -190,7 +190,7 @@
   category: Penis
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -218,7 +218,7 @@
   category: Penis
   userArouse: 5
   targetArouse: 20
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bj1.ogg
   - /Audio/_Arcane/Interactions/bj2.ogg
@@ -253,7 +253,7 @@
   category: Butt
   userArouse: 10
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/friction1.ogg
   - /Audio/_Arcane/Interactions/friction2.ogg
@@ -285,7 +285,7 @@
   category: Mouth
   userArouse: 15
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bj1.ogg
   - /Audio/_Arcane/Interactions/bj2.ogg
@@ -320,7 +320,7 @@
   category: Penis
   userArouse: 5
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/friction1.ogg
   - /Audio/_Arcane/Interactions/friction2.ogg
@@ -352,7 +352,7 @@
   id: HornsBlowjobF
   category: Head
   targetArouse: 10
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/oral1.ogg
   - /Audio/_Arcane/Interactions/oral2.ogg
