+++ Resources/Prototypes/_Arcane/Interactions/ERP/pussy.yml
@@ -97,7 +97,7 @@
   id: PussyTouch
   category: Pussy
   targetArouse: 5
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -122,7 +122,7 @@
   category: Pussy
   userArouse: 5
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
@@ -156,7 +156,7 @@
   category: Pussy
   userArouse: 15
   targetArouse: 20
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/bang1.ogg
   - /Audio/_Arcane/Interactions/bang2.ogg
@@ -192,7 +192,7 @@
   id: ToyPussy
   category: Pussy
   targetArouse: 15
-  range: 1
+  range: 1.5
   sounds:
   - /Audio/_Arcane/Interactions/champ_fingering.ogg
   messages:
