+++ Resources/Prototypes/_Arcane/Shaders/shaders.yml
@@ -1,0 +1,4 @@
+- type: shader
+  id: ArousalScreenEffect
+  kind: source
+  path: "/Textures/_Arcane/Shaders/arousal_screen_effect.swsl"
